Updates Go to 1.21, golangci-lint to 1.55.2
Note that, due to the change in import and init func ordering changing
in Go 1.21, the grpc proxy proto codec is no longer registered after the
grpc internal proto codec definition, meaning the dapr custom definition
will not be registered and will cause errors. The custom proto needs to
be either re-registered at runtime, or change its name in order to fix
this. This is done in the gRPC object, and handler_test.go TestMain.

Signed-off-by: joshvanl <me@joshvanl.dev>